Feat/better intake tests (#12)
Adds better intake unit tests. Nearly all functionality is now tested and passes.

* adds intake power unit test per #10

* adds indexer power unit test per #10

* added intake position unit test per #10 (this was the really long one to test the deploy/retract logic)
to help this, also made some getters for the intake direction and pid setpoint internal vars. we might even use those in a command later on, so hey, less work for later

* added intake get position unit test per #10
this also meant adding IsExtended() and IsRetracted() methods to the intake, per the comment on #10

* fixed an inordinate amount of basic syntax errors

* the Intake Init() method now updates the entire config, so hopefully that should actually add values to the intake config and so the unit tests should then pass

* fixed intake position min output and unit test not working because the intake hand't been commanded to move
at this point I'm not sure if the IntakeGetPositionTest is useful at all, because the IntakePositionTest is quite thorough and covers basically the same thing

* delete the IntakeGetPositionTest because it overlapped too much with IntakePositionTest
also it wasn't working and I don't know why

* fixes unit test because it relied on a part of the HardwareInterface code, which obviously wouldn't be run in a unit test like this because we don't have hardware

* fixes IsExtended not using inverse logic

* fixed unit tests not running on my machine
also tried to fix the intake position test case. some of it is probably due to floating point randomness, which is why the IsRetracted now has a margin of error
but now changing something on line 91 makes a check on line 81 fail so I have no idea what's going on

* added VOKC_CALL around clamp call in Intake periodic

* fix intake unit tests per #10
added some debug couts that need to get removed before merging
also tests aren't completely fixed because of random -6.27744e66 that keeps showing up
just committing and pushing this so justin can see it

* fix comparison on line 136 of intake position test
if a number is less negative then it is more positive, hence greater than

* initialize the intake position PID to 0 so we don't have an uninitialized  variable laying around

* fixes intake position unit test per #10
just had to add intake_->Init() in the test, which means that things do not carry over from test to test

* removed cout debug statements prior to merge with master as unit tests now all pass fine

* added Init() in the SetUp() function for the intake unit tests so everything gets initialized to a reasonable default value